pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ad-sync"
version = "0.1.0"
description = "Automatically sync audio descriptions from AudioVault with Sonarr/Radarr downloads"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "describealign>=2.0",
]

[project.scripts]
ad-sync = "ad_sync.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ad_sync*"]
